streaming

+++ src/main/scala/Tables.scala
@@ -24,15 +24,15 @@ object Suppliers{
 
 // A Coffees table with 5 columns: name, supplier id, price, sales, total
 class Coffees(tag: Tag)
-  extends Table[(String, Int, Double, Int, Int)](tag, "COFFEES") {
+  extends Table[(String, Int, Float, Int, Int)](tag, "COFFEES") {
 
   def name: Rep[String] = column[String]("COF_NAME", O.PrimaryKey)
   def supID: Rep[Int] = column[Int]("SUP_ID")
-  def price: Rep[Double] = column[Double]("PRICE")
+  def price: Rep[Float] = column[Float]("PRICE")
   def sales: Rep[Int] = column[Int]("SALES")
   def total: Rep[Int] = column[Int]("TOTAL")
   
-  def * : ProvenShape[(String, Int, Double, Int, Int)] =
+  def * : ProvenShape[(String, Int, Float, Int, Int)] =
     (name, supID, price, sales, total)
   
   // A reified foreign key relation that can be navigated to create a join
@@ -41,5 +41,5 @@ class Coffees(tag: Tag)
 }
 
 object Coffees {
-  type Row = (String, Int, Double, Int, Int)
+  type Row = (String, Int, Float, Int, Int)
 }
